If useLocal actually use localfs
In some cases it is desirable to add files to the local fs and test without a need to rebuild an executable.  Previously _ascLocalFS.Open() implementation needed the path to exist in the generated map of files in the fs or would fail.  This change removes the dependency for the path to already exist in static FS generation.

Eventually It would be nice to change FS to take a string argument that would direct which filesystem path to use instead of defaulting to to predefined local path.  so after deployment of binary an argument could be provided to allow user to specify a path to use for content.